requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
fastapi==0.115.14
greenlet==3.2.3
grpcio==1.73.0
grpcio-tools==1.73.0
h11==0.16.0
idna==3.10
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
redis>=4.0.0
requests==2.32.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
